requirements.txt
pandas
psycopg2-binary
streamlit
openpyxl
dotenv
matplotlib
sqlalchemy
tabulate
numpy